+++ R/klass.R
@@ -8,15 +8,15 @@
 #' both municipality and county codes in the same function call. All three
 #' functions default to the current year.
 #'
-#' The functions returns a data.frame by default. If you prefer the the output
+#' The functions returns a data.frame by default. If you prefer the output
 #' as a list you can set `simplify` to `FALSE`. This can be useful to keep
 #' programmatically track of failed queries. If you set `raw_response`
 #' to `TRUE`, the raw response from the API will be returned together with the
 #' parsed response. Note that the response will then be returned silently.
 #'
 #' @inheritParams get_entity
-#' @param year The year for which the codes should be valid for
-#' @param include_notes If notes should be included or not
+#' @param year The year for which the codes should be valid for.
+#' @param include_notes If notes should be included or not.
 #'
 #' @return data.frame or list
 #'
@@ -37,7 +37,7 @@ get_municipalities <- function(
 
   common_info(simplify, raw_response)
 
-  dl <- lapply(year, get_klass_codes_single, type = 'municipality',
+  dl <- lapply(year, get_klass_codes_safe, type = 'municipality',
                include_notes = include_notes, raw_response = raw_response)
   if (raw_response) return(dl)
   if (simplify) return(do.call('rbind', dl))
@@ -50,7 +50,7 @@ get_counties <- function(
     year = format(Sys.Date(), '%Y'), include_notes = FALSE, simplify = TRUE,
     raw_response = FALSE) {
   common_info(simplify, raw_response)
-  dl <- lapply(year, get_klass_codes_single, type = 'county',
+  dl <- lapply(year, get_klass_codes_safe, type = 'county',
                include_notes = include_notes, raw_response = raw_response)
   if (raw_response) return(dl)
   if (simplify) return(do.call('rbind', dl))
@@ -59,10 +59,9 @@ get_counties <- function(
 
 #' @rdname get_municipalities
 #' @export
-get_adm_units <- function(
-    year = format(Sys.Date(), '%Y'), include_notes = FALSE, simplify = TRUE) {
-  dl <- lapply(year, get_klass_codes_single, type = 'both',
-               include_notes = include_notes, raw_response = FALSE)
+get_adm_units <- function(year = format(Sys.Date(), '%Y'), simplify = TRUE) {
+  dl <- lapply(year, get_klass_codes_safe, type = 'both',
+               include_notes = FALSE, raw_response = FALSE)
   if (simplify) return(do.call('rbind', dl))
   dl
 }
@@ -80,13 +79,13 @@ get_adm_units <- function(
 #' All years from 1974 until present are supported. The function defaults to
 #' the current year.
 #'
-#' If notes are enables with `include_notes`, a column `note` will be added to the
+#' If notes are enabled with `include_notes`, a column `note` will be added to the
 #' data.frame. Notes are stated with a code reference that has the following order
 #' (alpha-2, alpha-3, num-3, SSB-3), for example Norway (NO, NOR, 578, 000). See
 #' [Statistics Norway's webpage](https://www.ssb.no/klass/klassifikasjoner/552) for
 #' more information on the code specification.
 #'
-#' The function returns a data.frame by default. If you prefer the the output
+#' The function returns a data.frame by default. If you prefer the output
 #' as a list you can set `simplify` to `FALSE`. This can be useful to keep
 #' programmatically track of failed queries. If you set `raw_response`
 #' to `TRUE`, the raw response from the API will be returned together with the
@@ -104,7 +103,46 @@ get_countries <- function(
     year = format(Sys.Date(), '%Y'), include_notes = FALSE, simplify = TRUE,
     raw_response = FALSE) {
   common_info(simplify, raw_response)
-  dl <- lapply(year, get_klass_codes_single, type = 'country',
+  dl <- lapply(year, get_klass_codes_safe, type = 'country',
+               include_notes = include_notes, raw_response = raw_response)
+  if (raw_response) return(dl)
+  if (simplify) return(do.call('rbind', dl))
+  dl
+}
+
+#' Get Standard Industrial Classification (SIC)
+#'
+#' Get all Standard Industrial Classification (SIC) based on the NACE standard.
+#'
+#' Standard Industrial Classification (SIC) is primarily a statistical standard
+#' and is the basis for coding units according to the most important activities
+#' in Statistics Norway's Business register and in the Central Coordinating
+#' Register for Legal Entities. SIC2007 is based on NACE Rev. 2. See
+#' [Statistics Norway's webpage](https://www.ssb.no/klass/klassifikasjoner/6) for
+#' more information on the code specification.
+#'
+#' All years from 1994 until present are supported. The function defaults to the
+#' current year. If notes are enabled with `include_notes`, a column `note` will
+#' be added to the data.frame.
+#'
+#' The function returns a data.frame by default. If you prefer the output
+#' as a list you can set `simplify` to `FALSE`. This can be useful to keep
+#' programmatically track of failed queries. If you set `raw_response`
+#' to `TRUE`, the raw response from the API will be returned together with the
+#' parsed response. Note that the response will then be returned silently.
+#'
+#' @inheritParams get_municipalities
+#'
+#' @return data.frame or list
+#'
+#' @seealso [get_entity()]
+#'
+#' @export
+get_industrial_codes <- function(
+    year = format(Sys.Date(), '%Y'), include_notes = FALSE, simplify = TRUE,
+    raw_response = FALSE) {
+  common_info(simplify, raw_response)
+  dl <- lapply(year, get_klass_codes_safe, type = 'sic',
                include_notes = include_notes, raw_response = raw_response)
   if (raw_response) return(dl)
   if (simplify) return(do.call('rbind', dl))
@@ -114,7 +152,7 @@ get_countries <- function(
 #' get_klass_codes_single
 #' @noRd
 get_klass_codes_single <- function(
-    year, type = c('municipality', 'county', 'country', 'both'),
+    year, type = c('municipality', 'county', 'country', 'sic', 'both'),
     include_notes = FALSE, raw_response = FALSE)
 {
   if (type == 'both') {
@@ -130,7 +168,8 @@ get_klass_codes_single <- function(
     type,
     'municipality' = klass_municipalities,
     'county' = klass_counties,
-    'country' = klass_countries)
+    'country' = klass_countries,
+    'sic' = klass_sic)
   endpoint <- get_klass_api_endpoint(
     year, type = type, links = x$endpoint, from = x$valid_from, to = x$valid_to
   )
@@ -143,3 +182,10 @@ get_klass_codes_single <- function(
   }
   parsed
 }
+
+#' Get klass (safe method)
+#' @inheritParams get_klass_codes_single
+#' @noRd
+get_klass_codes_safe <-
+  purrr::possibly(get_klass_codes_single, otherwise = NULL, quiet = FALSE)
+