+++ tests/testthat/test-klass.R
@@ -1,7 +1,32 @@
 # ---- Test for warnings and errors ----
 test_that('get_municipality() issue warnings and errors as expected', {
-  expect_error(get_municipalities(year = 1776))
-  expect_error(get_municipalities('abc'))
+  expect_message(expect_null(get_municipalities(year = 1776)))
+  expect_message(expect_null(get_municipalities(year = 17776)))
+  expect_message(expect_null(get_municipalities('abc')))
+})
+
+test_that('get_countries() issue warnings and errors as expected', {
+  expect_message(expect_null(get_countries(year = 1776)))
+  expect_message(expect_null(get_countries(year = 17776)))
+  expect_message(expect_null(get_countries('abc')))
+})
+
+test_that('get_counties() issue warnings and errors as expected', {
+  expect_message(expect_null(get_counties(year = 1776)))
+  expect_message(expect_null(get_counties(year = 17776)))
+  expect_message(expect_null(get_counties('abc')))
+})
+
+test_that('get_adm_units() issue warnings and errors as expected', {
+  expect_message(expect_null(get_adm_units(year = 1776)))
+  expect_message(expect_null(get_adm_units(year = 17776)))
+  expect_message(expect_null(get_adm_units('abc')))
+})
+
+test_that('get_industrial_codes() issue warnings and errors as expected', {
+  expect_message(expect_null(get_industrial_codes(year = 1776)))
+  expect_message(expect_null(get_industrial_codes(year = 17776)))
+  expect_message(expect_null(get_industrial_codes('abc')))
 })
 
 # Skip rest of tests if live API test are disabled
@@ -34,7 +59,6 @@ test_that('get_municipality() returns expected results', {
 
   # --- Tests for raw response ---
   df <- suppressMessages(get_municipalities(raw_response = TRUE))
-  expect_true(is.list(df))
   expect_true(is.list(df))
   expect_true(is.list(df[[1]]))
   expect_equal(names(df[[1]]), c('url', 'status', 'content', 'response'))
@@ -138,4 +162,32 @@ test_that('get_counties() returns expected results', {
   expect_equal(names(df[[1]]), c('url', 'status', 'content', 'response'))
   expect_identical(class(df[[1]]), "noAPI")
   expect_true(is.list(df[[1]]$content))
+})
+
+# ---- get_klass_codes_single() ----
+
+test_that('get_klass_codes_single() fails correctly', {
+  expect_error(get_klass_codes_single(1789, 'municipality'))
+  expect_error(get_klass_codes_single('abc', 'municipality'))
+})
+
+# ---- get_industrial_codes() ----
+test_that('get_industrial_codes() returns expected results', {
+  skip_on_cran()
+  skip_if(check_api(ssb_url))
+
+  df <- get_industrial_codes()
+  expect_equal(class(df), 'data.frame')
+  expect_equal(colnames(df), c('year', 'code', 'parentCode', 'level', 'shortName', 'name'))
+
+  # --- Tests for list ---
+  df <- get_industrial_codes(simplify = FALSE, include_notes = TRUE)
+  expect_true(is.list(df))
+  expect_equal(class(df[[1]]), 'data.frame')
+
+  # --- Tests for raw response ---
+  df <- get_industrial_codes(raw_response = TRUE)
+  expect_true(is.list(df))
+  expect_true(is.list(df[[1]]))
+  expect_identical(class(df[[1]]), "noAPI")
 })