+++ tests/testthat/test-utils.R
@@ -1,17 +1,21 @@
 skip_if(Sys.getenv('NOAPI_SKIP_LIVE_API_TESTS') == 'TRUE')
 
 test_that('check_api() works', {
+  skip_if(check_api('https://httpbin.org'))
   skip_on_cran()
+  skip_on_ci()
+
   expect_false(check_api('http://httpbin.org/status/200'))
   expect_true(check_api('http://httpbin.org/status/404'))
   expect_true(check_api('http://httpbin.org/status/500'))
 })
 
 test_that('send_query() retries on specific error codes', {
+  skip_if(check_api('https://httpbin.org'))
   skip_on_cran()
   skip_on_ci()
 
-    # 400 (no retry)
+  # 400 (no retry)
   u <- request('http://httpbin.org/status/400')
   tmp1 <- bench::system_time(send_query(u, max_tries = 3, throttle_rate = 1))
   # 429 (should retry)
@@ -25,7 +29,9 @@ test_that('send_query() retries on specific error codes', {
 })
 
 test_that('parse_response() works for JSON-formats', {
+  skip_if(check_api('https://httpbin.org'))
   skip_on_cran()
+  skip_on_ci()
 
   u <- 'https://httpbin.org/json'
   resp <- request(u) |> req_perform()
@@ -40,12 +46,17 @@ test_that('parse_response() works for JSON-formats', {
 })
 
 test_that('parse_response() returns error for unknown formats', {
+  skip_if(check_api('https://httpbin.org'))
+  skip_on_cran()
+  skip_on_ci()
+
   u <- 'https://httpbin.org/html'
   resp <- request(u) |> req_perform()
   expect_error(parse_response(resp))
 })
 
 test_that('make_api_object() works', {
+  skip_if(check_api('https://httpbin.org'))
   skip_on_cran()
   skip_on_ci()