+++ NEWS.md
@@ -1,3 +1,33 @@
-# noAPI 0.0.0.9000
+# noAPI 0.1.0.9000
 
-* Added a `NEWS.md` file to track changes to the package.
+## New features
+
+* Add new function `get_schools()` (#33)
+
+
+# noAPI 0.1.0 (March 29 2023)
+
+This is the first release of noAPI, introducing the following functions:
+
+Brønnøysundsregisteret:
+
+* `get_entity()`
+* `get_roles()`
+
+Kartverket:
+
+* `get_address_info()`
+* `find_address_from_point()`
+
+SSB:
+
+* `get_adm_units()`
+* `get_municipalities()`
+* `get_counties()`
+* `get_countries()`
+* `get_industrial_codes()`
+
+Norges Bank:
+
+* `get_exchange_rate()`
+